# perf: switch to a faster xxhash package