### Why the changes in this PR are needed?

<!--
Include a short description of WHY the changes were made.
-->

```
go test -v -benchmem -bench '^BenchmarkTermHashing$' -run='^$' -count=10 github.com/open-policy-agent/opa/v1/ast

goos: linux
goarch: amd64
pkg: github.com/open-policy-agent/opa/v1/ast
cpu: AMD Ryzen 7 PRO 4750U with Radeon Graphics
                    │   old.txt   │               new.txt               │
                    │   sec/op    │   sec/op     vs base                │
TermHashing/10-16     18.68n ± 2%   11.30n ± 0%  -39.49% (p=0.000 n=10)
TermHashing/100-16    42.94n ± 2%   33.71n ± 1%  -21.48% (p=0.000 n=10)
TermHashing/1000-16   179.4n ± 0%   165.1n ± 1%   -7.97% (p=0.000 n=10)
geomean               52.39n        39.77n       -24.10%
```

### What are the changes in this PR?

<!--
Include a short description of WHAT changes were made.
-->

Replace [`github.com/OneOfOne/xxhash`](https://github.com/OneOfOne/xxhash) with [`github.com/cespare/xxhash/v2`](https://github.com/cespare/xxhash), a faster implementation of xxHash.


### Notes to assist PR review:

<!--
Here you can add information you think will help the reviewer(s).
-->

### Further comments:

<!--
Here you can include links to additional resources related to the changes, discuss your solution, other approaches you considered etc.
-->